[454] Ядро / Предметы: Исправлены КД предметов

+++ src/server/game/Entities/Player/Player.h
@@ -330,7 +330,7 @@ struct CUFProfile
 struct SpellCooldown
 {
     time_t end;
-    uint16 itemid;
+    uint32 itemid;
 };
 
 typedef std::map<uint32, SpellCooldown> SpellCooldowns;
